[COMMUNITY] gles-user-module: fix dependencies
Users of gles-user-module can hit link issues since it does not have
libgbm and wayland-kms in its DEPENDS, but they are required to link
against the libraries it provides.  To fix this, add them to DEPENDS
per OpenEmbedded best practices.  Doing so requires spliting the EGL
headers into a separate recipe to avoid a circular dependency problem
with wayland-kms.

Bug-AGL: SPEC-429

Change-Id: If7c678b1dd0b222074dfb81823874f48ce790d88
Signed-off-by: Scott Murray <scott.murray@konsulko.com>